backend/requirements.txt
# requirements.txt

asgiref==3.8.1
captcha==0.6.0
channels==4.2.0
click==8.1.7
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
h11==0.14.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.2
uvicorn==0.32.1
